Update inline with how laravel's mix function works
This adds an optional $base property which allows you to set the relative path for where the assets are to be found. Say you have your assets build into `/assets/build`. The manifest would have mappings relative to `/assets/build` and so the asset function would return `js/app.js` or `css/main.css`, etc. With the $base you can set it so it maps correctly to `/asset/build/js/app.js`, etc.